src/dcdfg_wrapper/requirements.txt
causaldag==0.1a163
cdt==0.5.23
conditional_independence==0.1a5
networkx==2.6.3
numba==0.55.0
numpy==1.21.5
pandas==1.4.0
pytorch_lightning==1.5.10
scanpy==1.9.1
scikit_learn==1.1.2
setuptools==59.5.0
tqdm==4.62.3
wandb==0.12.11
